Make the flex children container private
Reverse modifier application order
